fix: change to long

+++ src/main/java/com/boatarde/regatasimulator/util/FileUtils.java
@@ -46,7 +46,7 @@ public class FileUtils {
      * @return a list of Paths to the generated zip files.
      * @throws IOException if an I/O error occurs.
      */
-    public static List<Path> zipInChunks(String sourceDirPath, double chunkSize) throws IOException {
+    public static List<Path> zipInChunks(String sourceDirPath, long chunkSize) throws IOException {
         List<Path> zipFiles = new ArrayList<>();
         Path templatesPath = Paths.get(sourceDirPath);
